nillable deploy resource + nillable LWRP args
introduced nillable_set_or_return so that we can actually set a nil value
on a resource from the DSL.

this fixes the case where you want to do (and we document) setting
symlink_before_migrate to nil on the deploy resource.

to implement this better also moved up the attribute DSL method to the
base Chef::Resource class.

all LWRP resources are now nillable as well.